+++ src/pages/Credits/index.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Simple Weather - credits</title>
+    <link rel="stylesheet" href="/src/" />
+    <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
+  </head>
+  <body>
+    <main>
+      <h1>Credits</h1>
+      <h2>This app was created using the following resources:</h2>
+      <h3>Weather Icons</h3>
+      <p>
+        <a href="https://www.figma.com/community/file/1126777451931792118">Weather Glassmorphism Icons</a> by:
+        <a href="https://www.figma.com/@arve">Bilal Arve</a>
+      </p>
+      <h3>Favicon</h3>
+      <p>
+        <a href="https://www.svgrepo.com/svg/315013/weather">Glassy Ui Vectors</a> by:
+        <a href="https://dribbble.com/Spline">Spline</a>
+      </p>
+      <h3>Design</h3>
+      <p>
+        <a href="https://www.figma.com/community/file/1005161083412880387">Weather App</a> by:
+        <a href="https://www.figma.com/@mayankchaudhary">Mayank Chaudhary</a>
+      </p>
+      <h3>APIs</h3>
+      <p>
+        <a href="https://openweathermap.org/api">OpenWeatherMap</a> and
+        <a href="https://open-meteo.com/">Open-meteo</a>
+      </p>
+    </main>
+  </body>
+</html>
