+++ src/pages/Home/index.html
@@ -25,8 +25,8 @@
       name="description"
       content="Simplest weather, the sleek and modern weather web app, designed with a minimalist glass aesthetic that shows concise and simple information on the current weather conditions and forecast for the next few days."
     />
-    <link rel="stylesheet" href="style.css" />
-    <link rel="shortcut icon" href="img/weather-icon.svg" type="image/x-icon" />
+    <link rel="stylesheet" href="./styles.css" />
+    <link rel="shortcut icon" href="../../assets/img/weather-icon.svg" type="image/x-icon" />
   </head>
   <body class="morning-clrs">
     <main class="app-container">
@@ -41,7 +41,7 @@
       </header>
 
       <section class="weather-section">
-        <img src="img/weather-icons/01d.svg" alt="" />
+        <img src="../../assets/img/weather-icons/01d.svg" alt="" />
         <div class="temp-wrapper">
           <div class="temp">
             <h1>00</h1>
@@ -60,23 +60,23 @@
 
       <section class="weather-details">
         <div class="detail-wind detail">
-          <img src="img/wind-direction-icon.svg" alt="" />
+          <img src="../../assets/img/wind-direction-icon.svg" alt="" />
           <p>Wind speed: <span class="wind">00Km/h</span></p>
         </div>
         <span>|</span>
         <div class="detail-humity detail">
-          <img src="img/hum-icon.svg" alt="" />
+          <img src="../../assets/img/hum-icon.svg" alt="" />
           <p>Humidity: <span class="hum">00%</span></p>
         </div>
         <span>|</span>
         <div class="detail-rain detail">
-          <img src="img/rain-icon.svg" alt="" />
+          <img src="../../assets/img/rain-icon.svg" alt="" />
           <p>Cloudness: <span class="clouds">0.0%</span></p>
         </div>
       </section>
 
       <section class="forecast-section">
-        <img class="arrow-left" src="img/arrow-right.svg" alt="" />
+        <img class="arrow-left" src="../../assets/img/arrow-right.svg" alt="" />
         <section class="forecast">
           <!-- <div class="forecast-day-card">
             <p>00</p>
@@ -84,7 +84,7 @@
             <p class="week-day">Sun</p>
           </div> -->
         </section>
-        <img class="arrow-right" src="img/arrow-right.svg" alt="" />
+        <img class="arrow-right" src="../../assets/img/arrow-right.svg" alt="" />
       </section>
     </main>
 
@@ -106,6 +106,6 @@
       </nav>
     </footer>
 
-    <script type="module" src="app.js"></script>
+    <script type="module" src="main.js"></script>
   </body>
 </html>
