--- texts.js
@@ -1,6 +1,0 @@
-export const TEXT = {
-  error: {
-    TEXT_NO_GEOLOCATION_BROWSER_SUPPORT:
-      'Your browser does not support geolocation.',
-  },
-};
